Move container components to Routes directory
