+++ platform/viewer/src/routes/StudyListContainer/StudyListContainer.jsx
